[book2] Section 06-01, 02

+++ app/controllers/admin/staff_members_controller.rb
@@ -48,8 +48,12 @@ class Admin::StaffMembersController < Admin::Base
 
   def destroy
     staff_member = StaffMember.find(params[:id])
-    staff_member.destroy!
-    flash.notice = "職員アカウントを削除しました。"
+    if staff_member.deletable?
+      staff_member.destroy!
+      flash.notice = "職員アカウントを削除しました。"
+    else
+      flash.alert = "この職員アカウントは削除できません。"
+    end
     redirect_to :admin_staff_members
   end
 end
